Deployment/BLSTM/pytorch_server/requirements.txt
charset-normalizer==2.0.12
click==8.1.3
filelock==3.6.0
huggingface-hub==0.5.1
idna==3.3
joblib==1.1.0
packaging==21.3
pillow==9.1.0
pyparsing==3.0.8
pyyaml==6.0
regex==2022.4.24
requests==2.27.1
sacremoses==0.0.53
tensorflow-cpu==2.7.0
tokenizers==0.12.1
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
transformers==4.18.0
typing-extensions>=3.7.4
urllib3==1.26.9
numpy>=1.19.2
pandas>=1.4.2